Update 355873534_abilities.txt
Added localization for turrets and outposts - and the now-removed Counterintelligence ability. Also changed the description for Go Down Fighting - I misunderstood the purpose of the 'Hits' variable, causing me to misuse it. As a result of this, references to 'Hits' in GDF and the Flak Cannon will be changed to be in line with torpedoes instead of scaling with power/size.